BackEnd/requirements.txt
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pyserial==3.5
pytest==9.0.2
